projects/token-tracker/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "token-tracker"
version = "0.1.0"
description = "FastAPI service with LiteLLM telemetry and chat routes."
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "litellm>=1.44.7",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "boto3>=1.35.0",
]

[project.scripts]
token-tracker = "token_tracker.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["token_tracker*"]
